Add missing json scalar list tests. Separate tests.
