+++ query-engine/connector-test-kit-rs/query-engine-tests/tests/writes/data_types/mod.rs
@@ -4,4 +4,4 @@ mod datetime;
 mod decimal;
 mod json;
 mod native_types;
-mod scalar_lists;
+mod scalar_list;
