+++ query-engine/connector-test-kit-rs/query-engine-tests/tests/writes/data_types/scalar_list/base.rs
@@ -1,7 +1,7 @@
 use query_engine_tests::*;
 
 #[test_suite(schema(schema), capabilities(ScalarLists))]
-mod scalar_lists {
+mod basic_types {
     use indoc::indoc;
     use query_engine_tests::{assert_error, run_query, TROUBLE_CHARS};
 
@@ -21,7 +21,8 @@ mod scalar_lists {
             enum MyEnum {
               A
               B
-            }"#
+            }
+          "#
         };
 
         schema.to_owned()
@@ -300,215 +301,5 @@ mod scalar_lists {
             .await?
             .assert_success();
         Ok(())
-    }
-}
-
-#[test_suite(schema(schema), capabilities(ScalarLists, DecimalType))]
-mod dec_scalar_lists {
-    use indoc::indoc;
-    use query_engine_tests::run_query;
-
-    fn schema() -> String {
-        let schema = indoc! {
-            r#"model ScalarModel {
-              #id(id, Int, @id)
-              decimals  Decimal[]
-            }"#
-        };
-
-        schema.to_owned()
-    }
-
-    // "Scalar lists" should "be behave like regular values for create and update operations"
-    // Skipped for CockroachDB, lools like this is concat is also broken.
-    #[connector_test(exclude(CockroachDb))]
-    async fn behave_like_regular_val_for_create_and_update(runner: Runner) -> TestResult<()> {
-        insta::assert_snapshot!(
-          run_query!(&runner, format!(r#"mutation {{
-            createOneScalarModel(data: {{
-              id: 1,
-              decimals:  {{ set: ["1.234", "1.45"] }}
-            }}) {{
-              decimals
-            }}
-          }}"#, )),
-          @r###"{"data":{"createOneScalarModel":{"decimals":["1.234","1.45"]}}}"###
-        );
-
-        insta::assert_snapshot!(
-          run_query!(&runner, r#"mutation {
-            updateOneScalarModel(where: { id: 1 }, data: {
-              decimals:  { set: ["1.2345678"] }
-            }) {
-              decimals
-            }
-          }"#),
-          @r###"{"data":{"updateOneScalarModel":{"decimals":["1.2345678"]}}}"###
-        );
-
-        insta::assert_snapshot!(
-          run_query!(&runner, r#"mutation {
-            updateOneScalarModel(where: { id: 1 }, data: {
-              decimals:  { push: "2" }
-            }) {
-              decimals
-            }
-          }"#),
-          @r###"{"data":{"updateOneScalarModel":{"decimals":["1.2345678","2"]}}}"###
-        );
-
-        insta::assert_snapshot!(
-          run_query!(&runner, r#"mutation {
-            updateOneScalarModel(where: { id: 1 }, data: {
-              decimals:  { push: ["3", "4"] }
-            }) {
-              decimals
-            }
-          }"#),
-          @r###"{"data":{"updateOneScalarModel":{"decimals":["1.2345678","2","3","4"]}}}"###
-        );
-
-        Ok(())
-    }
-
-    // "A Create Mutation" should "create and return items with list values with shorthand notation"
-    #[connector_test]
-    async fn create_mut_work_with_list_vals(runner: Runner) -> TestResult<()> {
-        insta::assert_snapshot!(
-          run_query!(&runner, r#"mutation {
-            createOneScalarModel(data: {
-              id: 1
-              decimals: ["1.234", "1.45"]
-            }) {
-              decimals
-            }
-          }"#),
-          @r###"{"data":{"createOneScalarModel":{"decimals":["1.234","1.45"]}}}"###
-        );
-
-        Ok(())
-    }
-
-    // "A Create Mutation" should "create and return items with empty list values"
-    #[connector_test]
-    async fn create_mut_return_items_with_empty_lists(runner: Runner) -> TestResult<()> {
-        insta::assert_snapshot!(
-          run_query!(&runner, r#"mutation {
-            createOneScalarModel(data: {
-              id: 1
-              decimals: []
-            }) {
-              decimals
-            }
-          }"#),
-          @r###"{"data":{"createOneScalarModel":{"decimals":[]}}}"###
-        );
-
-        Ok(())
-    }
-
-    // "An Update Mutation that pushes to some empty scalar lists" should "work"
-    // Skipped for CockroachDB as enum array concatenation is not supported (https://github.com/cockroachdb/cockroach/issues/71388).
-    #[connector_test(exclude(CockroachDb))]
-    async fn update_mut_push_empty_scalar_list(runner: Runner) -> TestResult<()> {
-        create_row(&runner, r#"{ id: 1 }"#).await?;
-        create_row(&runner, r#"{ id: 2 }"#).await?;
-
-        insta::assert_snapshot!(
-          run_query!(&runner, r#"mutation {
-            updateOneScalarModel(where: { id: 1 }, data: {
-              decimals:  { push: "2" }
-            }) {
-              decimals
-            }
-          }"#),
-          @r###"{"data":{"updateOneScalarModel":{"decimals":["2"]}}}"###
-        );
-
-        insta::assert_snapshot!(
-          run_query!(&runner, r#"mutation {
-            updateOneScalarModel(where: { id: 2 }, data: {
-              decimals:  { push: ["1", "2"] }
-            }) {
-              decimals
-            }
-          }"#),
-          @r###"{"data":{"updateOneScalarModel":{"decimals":["1","2"]}}}"###
-        );
-
-        Ok(())
-    }
-
-    async fn create_row(runner: &Runner, data: &str) -> TestResult<()> {
-        runner
-            .query(format!("mutation {{ createOneScalarModel(data: {}) {{ id }} }}", data))
-            .await?
-            .assert_success();
-        Ok(())
-    }
-}
-
-// General defaults tests
-#[test_suite(schema(common), capabilities(ScalarLists))]
-mod defaults {
-    use indoc::indoc;
-    use query_engine_tests::run_query;
-
-    // empty defaults
-
-    fn common() -> String {
-        let schema = indoc! {
-            r#"model ScalarModel {
-              #id(id, Int, @id)
-              strings   String[]   @default(["Foo", "Bar", "Baz"])
-              ints      Int[]      @default([1, 2, 3])
-              floats    Float[]    @default([1.1, 2.2, 3.3])
-              booleans  Boolean[]  @default([true, false, false, true])
-              enums     MyEnum[]   @default([A, B, B, A])
-              dateTimes DateTime[] @default(["2019-07-31T23:59:01.000Z", "2012-07-31T23:59:01.000Z"])
-              bytes     Bytes[]    @default(["dGVzdA==", "dA=="])
-            }
-
-            enum MyEnum {
-              A
-              B
-            }
-            "#
-        };
-
-        schema.to_owned()
-    }
-
-    #[connector_test]
-    async fn basic_write(runner: Runner) -> TestResult<()> {
-        insta::assert_snapshot!(
-          run_query!(&runner, r#"mutation {
-          createOneScalarModel(data: {
-            id: 1
-          }) {
-            strings
-            ints
-            floats
-            booleans
-            enums
-            dateTimes
-            bytes
-          }
-        }"#),
-          @r###"{"data":{"createOneScalarModel":{"strings":["Foo","Bar","Baz"],"ints":[1,2,3],"floats":[1.1,2.2,3.3],"booleans":[true,false,false,true],"enums":["A","B","B","A"],"dateTimes":["2019-07-31T23:59:01.000Z","2012-07-31T23:59:01.000Z"],"bytes":["dGVzdA==", "dA=="]}}}"###
-        );
-
-        Ok(())
-    }
-
-    fn decimal() -> String {
-        let schema = indoc! {
-            r#"model ScalarModel {
-            #id(id, Int, @id)
-            dec Decimal[]
-          }"#
-        };
-
-        schema.to_owned()
     }
 }
