+++ query-engine/connector-test-kit-rs/query-engine-tests/tests/writes/data_types/scalar_list/mod.rs
@@ -1,0 +1,4 @@
+mod base;
+mod decimal;
+mod defaults;
+mod json;
